+++ Packages/XMusica/Scripts/Editor/XM_UIStyleManager.cs
@@ -12,6 +12,7 @@ namespace XMusica.EditorUtilities {
 
         public static Texture2D vinstBinderWindowIcon;
         public static Texture2D vinstSampleWindowIcon;
+        public static Texture2D vinstTesterWindowIcon;
         public static Texture2D whiteKeyTexture;
 
         static XM_UIStyleManager() {
@@ -19,6 +20,7 @@ namespace XMusica.EditorUtilities {
 
             vinstBinderWindowIcon = AssetDatabase.LoadAssetAtPath(path + "/Editor Resources/Textures/vinst_binder_window.png", typeof(Texture2D)) as Texture2D;
             vinstSampleWindowIcon = AssetDatabase.LoadAssetAtPath(path + "/Editor Resources/Textures/vinst_sample_window.png", typeof(Texture2D)) as Texture2D;
+            vinstTesterWindowIcon = AssetDatabase.LoadAssetAtPath(path + "/Editor Resources/Textures/vinst_tester_window.png", typeof(Texture2D)) as Texture2D;
             whiteKeyTexture = AssetDatabase.LoadAssetAtPath(path + "/Editor Resources/Textures/key_white.png", typeof(Texture2D)) as Texture2D;
 
             whitePianoKey = new GUIStyle("button");