+++ Packages/XMusica/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.0] - 2023-07-15
+### Additions
+- This is the first official release of XMusica.
+- Added save and discard dialogs when the user would leave the binding window with unsaved changes.
+- Added a Virtual Instrument tester window.
+
 ## [0.9.0] - 2023-07-15
 ### Additions
 - This is the first experimental release of XMusica.